requirements.txt
streamlit
scikit-learn
numpy
pandas
pillow